Initial commit of mentor dashboard
Initial commit to get structure of dashboards in place.
Dashboard.tmpl.html is the parent of all dashboard pages and the
loadDashboard() function in the dashboard.js.coffee will load the
correct dashboard based on the role. Decided to not use ui-router for
this right now since it’s not necessary. Each new dashboard will have
its own controller and tmpl associated with it just like the mentor one.

#21