If/Else Statements to Fix Super Admin bug
Super Admin pages were breaking because they aren’t assigned to an org
and don’t have a name (which we are now calling for in the nav).

Added a basic if/else ruby statement to fix it for now. If there is a
better or ‘correct’ way to do it, please replace!