Progress Circle Animation & Color Tweaks
Sped up animation and tweaked the BG color to be neutral on dark and
light backgrounds for now.

Eventually it’d be great to do a semi-transparent grey circle, but it’s
above my mental capacity right now, and I don’t want to mess with Jakes
code too much!