Moving 'add' button up to each class
I absolutely understand this *might* confuse org admins in the case
that they need to add a student to a class that hasn’t been assigned
yet. But the positives of showing the add button for each class
outweighs the rare negative.

Eventually it’d be good to make the ‘add’ buttons in each class,
automatically assign the student to that class.