Hookup some numbers to OAD branch
Fixes #255
Fixes #254
Fixes #253

This implementation is really, really unacceptably sloppy, unoptimized,
and not ok to check into master - but it’s the first step to
understanding how these numbers should be hooked up (will refactor as
part of #259).

Numbers with an asterisk after them are hard coded strings (should just
be GPA and mentor escalation), everything else should be real .

The average statistics are on a per semester basis, but can easily be
changed to be out of the total lifetime of the organization.

Students that need attention are determined by a needs_attention
attribute on the student object, which is set based on expectations,
but can easily be changed to be based on other factors as well.

The biggest thing that has to be done next, is pull out the front end
parsing and calculations done on the ViewOrganizationWithUsers object,
which is commonly used across the org dashboard, mentors, and students
pages now. It would also be nice to optimize the calculations and data
pulled for each page, since right now they’re all doing the full extent
of calculations on pretty much all the data that an organization has…