Fixed failing User assignment tests
We need to revisit the get_assigned_students function because it is not
easy to test since it returns the full ViewOrganization object. We
either figure out a way to easily access all the properties in that
JSON object in ruby or break up the function into smaller functions so
we can test one of the smaller functions to make sure the relationships
array is being set correctly on the ViewUser object.

#325