Reverting back to old way without Bower...
This sucks but there were some weird issues with the assets compiling
when using Bower and Heroku and it wasn’t worth spending much more time
on it. So for now reverting back to using Gems for JS libs and directly
copy/pasting Angulartics.js files into vendor/assets/javascripts
(gross!).

This is another reason to move to a complete separation between front
and backend code so Heroku doesn’t deal with any assets or any front
end stuff

#358