Adding Labels to Tasks
@ttapers Another one I’ll need help with. I have an ‘Open’ and
‘Complete’ label (straight copy off of github haha). It’d also be great
to have a ‘Due Soon’ and ‘Past Due’ label on them. I’m not quite sure
how to access the date/time and compare it to the due date though.
There is also a case where there isn’t a due date, where the task
should just be ‘open’ or ‘complete’.