Added Default View
This may be a little hacky. If we end up liking the way it looks, it
wouldn’t hurt for @Jakeah to take another look at how the panes work.

The ‘Default’ view might be better as it’s own individual module that
loads in. I just don’t know how to do that.