!!!Font Change!!!
Be on the lookout for anything broken. Obviously, it’s a big thing to
change.

Originally the intention was to use Proxima Nova for the body font and
small headlines, and Facit for headlines. I think it’s time to give it
a shot. Proxima Nova has more contrast in it’s font-weights and I think
it looks better, unless we’re talking big headlines.