Changed how sessions are stored and set expiration to 15 minutes
Sessions are now stored in our DB instead of fully in a cookie on the
client’s computer. This is a little better for security and gives us
more control over forcing everyone to re-login by deleting the sessions.

Sessions expire every 15 minutes now and the user is redirected to the
login page after expiration.

Fixes #437