Adding Success Messages when classes are saved and deleted
With the new progress page design, things jump around a bit after
saving/deleting classes because of the size of the forms and the
accordion effect so having these messages appear will make it less
confusing as to what just happened. It’ll give the user a confirmation
of their action.

#528