Created a back-button attribute directive
The directive adds a click handler on the element which acts like the
browsers back button. I replaced all pages we currently have that had a
“Back to something” link to use this and they all just say “Go Back”
now so you’ll go back to wherever you came from.

Once the new design is finalized, this directive can be made into an
element directive with a template that spits out the html code so all
we need to do is put <back-button></back-button> wherever we want it

Fixes #575