Forgot to add current_user check in application_controller
One of the things we needed to bring back from #678. This situation is
hit when the user’s session is expired AND the app was just pushed.
Without the check for current_user, you get a 500 error because it
tries to do current_user.id which is trying to do null.id

#678