Removing session timeout for now
Problem has something to do with expired sessions and warden throwing
an Authentication failure when the session is expired. We want the app
to go to our Angular app where we’ll correctly redirect the user to the
login page with query string params. However, when session expires and
you refresh, the request stops when Warden throws an unauthenticated
error and goes to the failure app#respond which redirects to login page
but loses the query params. Need to find an expert in Devise and Warden
to help us understand this. For now, just getting rid of the session
timeout until we figure it out.

Fixes #694