Updated the slider text when showing only 1 semester
Also briefly changed the slider so it would only go as high as the
student’s current semester. Students who are in Semester 1 don’t see a
slider in this case. I felt this was actually more confusing to not see
a slider than it is to always see the slider even though it lets you go
out to Semester 8 for every kid (regardless of whether they have data
there or not)

#760