Moved all images into S3 and updated all URLS to point to our Amazon CloudFront distribution
This isn’t the cause of any memory issues but I’ve wanted to do this at
some point and figured I’d just get it over with now.

CloudFront is a CDN so our images will load much faster and won’t take
up space on our server. When the CDN doesn’t have an image cached on
one of its edge servers it’ll fall back to S3 for our images and then
cache that image.

#792